requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
python-dotenv==1.0.1
openai==1.30.1
python-multipart==0.0.9
